+++ src/main.ts
@@ -18,6 +18,7 @@ export {Debounce} from "./decorators/debounce.decorator";
 export {ElementAttribute} from "./decorators/element-attribute.decorator";
 export {EventListener} from "./decorators/event-listener.decorator";
 export {GlobalEventListener} from "./decorators/global-event-listener.decorator";
+export {WindowEventListener} from "./decorators/window-event-listener.decorator";
 /**
  * useful functions
  */