+++ CHANGELOG.md
@@ -8,7 +8,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ### Added
 
-- ...
+- WindowEventListener Decorator (works like the `GlobalEventListener`, but `window` is used instead of `document`)
 
 ## [3.2.0]